added validation for addCompany, updateCompany, and deletCompany

+++ controllers/companies.controllers.js
@@ -14,7 +14,12 @@ const addCompany = async (req, res) => {
     const { name, logoUrl, location } = req.body;
 
     if (!name || !logoUrl || !location) {
-      return res.status(400).json({ message: "Company name, logoURL, and location are required.", success: false });
+      return res
+        .status(400)
+        .json({
+          message: "Company name, logoURL, and location are required.",
+          success: false
+        });
     }
 
     await Companies.create({ name, logoUrl, location });
@@ -22,6 +27,7 @@ const addCompany = async (req, res) => {
     res.status(201).json({
       message: "New Company Created",
       success: true,
+      name: name, 
     });
   } catch (error) {
     res.status(500).json({ message: error.message, success: false });
@@ -31,26 +37,34 @@ const addCompany = async (req, res) => {
 const deleteComp = async (req, res) => {
   try {
     const { id } = req.params;
+
+    const company = await Companies.findById(id);
+
+    if (!company) {
+      return res.status(404).json({ message: "Company does not exist", success: false });
+    }
+
     await Companies.findByIdAndDelete(id);
 
     res.status(200).json({
-      message: "Company Deleted",
+      message: "Company Deleted:",
+      company,
       success: true,
     });
   } catch (error) {
-    res.status(500).json({ message: error.message, success: false });
+    res.status(400).json({ message: "Invalid company ID", success: false });
   }
 };
 
 const updateComp = async (req, res) => {
   try {
     const { id } = req.params;
-    const { name, logoUrl, location } = req.body.params;
+    const { name, logoUrl, location } = req.body;
 
-    const company = Company.findById(id);
+    const company =  await Companies.findById(id);
 
     if (!company) {
-      res.status(404).json({ message: "Company not found", success: false });
+      return res.status(404).json({ message: "Company not found", success: false });
     }
 
     const updated = await Companies.findByIdAndUpdate(
